Updated Character.
The player can now look at the mouse.

His movement has been updated so that he can move relative to how the
mesh is facing, how the camera is facing and how the world is facing. He
will also eventually move relative to how another object is facing. (To
be added.)

Also the lock on the free movement on the camera for the player has been
fixed. He should be able to freely rotation 360 degrees without any
clamping or juttering.

I have also created a Prefabs folder with a character waiting in it. I
also decoupled the input handling and the controller.